frontend/requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
mysql-connector-python>=8.4.0
python-dotenv>=1.0.0
petl==1.7.16